backend/requirements.txt
fastapi
uvicorn
langchain
pinecone-client
openai
PyPDF2
python-multipart
python-dotenv
